Refactored tests for move FRS tests to separate directories
