+++ test/connectors/VatRegistrationConnectorSpec.scala
@@ -237,27 +237,6 @@ class VatRegistrationConnectorSpec extends VatRegSpec with VatRegistrationFixtur
     }
   }
 
-  "Calling updateVatFlatRateScheme" should {
-
-
-    "return the correct VatResponse when the microservice completes and returns a VatFlatRateScheme model" in new Setup {
-      mockHttpPATCH[VatFlatRateScheme, VatFlatRateScheme]("tst-url", validVatFlatRateScheme)
-      connector.updateVatFlatRateScheme("tstID", validVatFlatRateScheme) returns validVatFlatRateScheme
-    }
-    "return the correct VatResponse when a Forbidden response is returned by the microservice" in new Setup {
-      mockHttpFailedPATCH[VatFlatRateScheme, VatFlatRateScheme]("tst-url", forbidden)
-      connector.updateVatFlatRateScheme("tstID", validVatFlatRateScheme) failedWith forbidden
-    }
-    "return a Not Found VatResponse when the microservice returns a NotFound response (No VatRegistration in database)" in new Setup {
-      mockHttpFailedPATCH[VatFlatRateScheme, VatFlatRateScheme]("tst-url", notFound)
-      connector.updateVatFlatRateScheme("tstID", validVatFlatRateScheme) failedWith notFound
-    }
-    "return the correct VatResponse when an Internal Server Error response is returned by the microservice" in new Setup {
-      mockHttpFailedPATCH[VatFlatRateScheme, VatFlatRateScheme]("tst-url", internalServiceException)
-      connector.updateVatFlatRateScheme("tstID", validVatFlatRateScheme) failedWith internalServiceException
-    }
-  }
-
   "Calling upsertLodgingOfficer" should {
 
     val vatLodgingOfficer = validLodgingOfficer
