+++ test/services/VatRegistrationServiceSpec.scala
@@ -278,26 +278,6 @@ class VatRegistrationServiceSpec extends VatRegSpec with VatRegistrationFixture 
       service.submitVatLodgingOfficer() failedWith classOf[IllegalStateException]
     }
 
-    "submitVatFlatRateScheme should process the submission even if VatScheme does not contain VatFlatRateScheme" in new Setup {
-      when(mockRegConnector.getRegistration(Matchers.eq(testRegId))(any(), any())).thenReturn(emptyVatScheme.pure)
-      when(mockRegConnector.upsertVatFlatRateScheme(any(), any())(any(), any())).thenReturn(validVatFlatRateScheme.pure)
-      save4laterReturns(S4LFlatRateScheme(
-        joinFrs = Some(JoinFrsView(true)),
-        annualCostsInclusive = Some(AnnualCostsInclusiveView("yes")),
-        annualCostsLimited = Some(AnnualCostsLimitedView("yes")),
-        registerForFrs = Some(RegisterForFrsView(true)),
-        frsStartDate = Some(FrsStartDateView(FrsStartDateView.VAT_REGISTRATION_DATE))
-      ))
-      service.submitVatFlatRateScheme() returns validVatFlatRateScheme
-    }
-
-    "submitVatFlatRateScheme should fail if there's no VatFlatRateScheme in backend or S4L" in new Setup {
-      when(mockRegConnector.getRegistration(Matchers.eq(testRegId))(any(), any())).thenReturn(emptyVatScheme.pure)
-      save4laterReturnsNothing[S4LFlatRateScheme]()
-
-      service.submitVatFlatRateScheme() failedWith classOf[IllegalStateException]
-    }
-
     "submitSicAndCompliance should fail if VatSicAndCompliance not in backend and S4L" in new Setup {
       when(mockRegConnector.getRegistration(Matchers.eq(testRegId))(any(), any())).thenReturn(emptyVatScheme.pure)
       save4laterReturnsNothing[S4LVatSicAndCompliance]()
